Move `ReplHost` to a new file
